+++ install.sh
@@ -66,6 +66,7 @@ function usageText {
   echo '                             gh - use authenticated Github commandline tool'
   echo '                             curl-authenticated - use curl authenticated with the token from the --token option'
   echo '                             curl - use curl without token - needs jq to parse rest response'
+  echo '                             archive=<archive> - use tar archive as release'
   echo "-s/--service-directory[$service_directory]"
   echo '                             directory for service files'
   echo "-c/--configuration-directory[$configuration_directory]"
@@ -139,6 +140,13 @@ while [[ $# -gt 0 ]]; do
         release_fetch_mode="curl"
       else
         usage '--release-fetch-mode set to curl but can'\''t find jq command, which is needed to parse the REST API'
+      fi
+      ;;
+    archive=*)
+      release_fetch_mode="archive"
+      archive_file="${value:8}"
+      if [ ! -f "$archive_file" ]; then
+        usage "couldn't find archive file '$archive_file'"
       fi
       ;;
     '')
@@ -256,11 +264,23 @@ function extractUrlFromGraphqlQuery() {
     fi
 }
 
+function downloadBinary() {
+  echo 'download binary' >&2
+  curl --progress-bar --header 'Accept: application/octet-stream' --url "${url}" --output "${archive_name}" --fail
+}
+
+archive_name='kss.tar.gz'
+
 case "$release_fetch_mode" in
+archive)
+  echo "use release archive from file $archive_file" >&2
+  cp "$archive_file" "$archive_name"
+  ;;
 gh)
   echo 'download latest release data via gh commandline tool' >&2
   response="$(gh api graphql -F 'user=EdwarDDay' -F 'repo=kotlin-server-scripts' -F 'asset=scripting-host-release.tar.gz' -f "query=$query")"
   url="$(extractUrlFromGraphqlQuery "$response")"
+  downloadBinary
   ;;
 curl-authenticated)
   echo 'download latest release data via curl and graphql api' >&2
@@ -274,18 +294,14 @@ curl-authenticated)
   }"
   response="$(curl --request POST "${github_args[@]}" --fail --silent --url 'https://api.github.com/graphql' --data "$data")"
   url="$(extractUrlFromGraphqlQuery "$response")"
+  downloadBinary
   ;;
 # curl
 *)
   echo 'download latest release data via github rest endpoint and jq' >&2
   url="$(curl --request GET "${github_args[@]}" --fail --silent --url 'https://api.github.com/repos/EdwarDDay/kotlin-server-scripts/releases/latest' | jq --raw-output '.assets | map(select(.content_type == "application/gzip"))[0].url')"
+  downloadBinary
 esac
-
-echo 'download binary' >&2
-
-archive_name='kss.tar.gz'
-
-curl --progress-bar --header 'Accept: application/octet-stream' --url "${url}" --output "${archive_name}" --fail
 
 echo 'extract binary' >&2